unit tests for 50% APY
