+++ common/src/venear.rs
@@ -60,3 +60,189 @@ impl VenearGrowthConfig {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use near_sdk::NearToken;
+
+    #[test]
+    fn test_50_percent_apy_linear_growth_calculation() {
+        // Test using the actual VenearGrowthConfig::calculate() function
+        let config = VenearGrowthConfig::FixedRate(Box::new(VenearGrowthConfigFixedRate {
+            annual_growth_rate_ns: Fraction {
+                numerator: 15854895991882.into(),  // 15,854,895,991,882
+                denominator: 10u128.pow(30).into(), // 10^30
+            },
+        }));
+        
+        let base_balance = NearToken::from_near(100);
+        let one_year_ns = 31_536_000_000_000_000u64;
+        
+        // Test 1 year growth
+        let start_time: TimestampNs = 0.into();
+        let end_time_1_year: TimestampNs = one_year_ns.into();
+        let growth_1_year = config.calculate(start_time, end_time_1_year, base_balance);
+        println!("Growth after 1 year: {} NEAR", growth_1_year.as_near());
+        
+        let actual_growth_1_year = growth_1_year.as_near();
+        println!("Actual growth after 1 year: {} NEAR", actual_growth_1_year);
+        
+        // Test 2 years growth
+        let end_time_2_years: TimestampNs = (one_year_ns * 2).into();
+        let growth_2_years = config.calculate(start_time, end_time_2_years, base_balance);
+        println!("Growth after 2 years: {} NEAR", growth_2_years.as_near());
+        
+        assert!(
+            actual_growth_1_year > 48 && actual_growth_1_year < 50,
+            "Growth should be positive and reasonable, got {} NEAR",
+            actual_growth_1_year
+        );
+        
+        // Test 4 years growth
+        let end_time_4_years: TimestampNs = (one_year_ns * 4).into();
+        let growth_4_years = config.calculate(start_time, end_time_4_years, base_balance);
+        println!("Growth after 4 years: {} NEAR", growth_4_years.as_near());
+        
+        println!("Current numerator: {}", 15854895991882u64);
+        println!("Growth results with numerator 15854895991882:");
+        println!("  - 1 year: {} NEAR growth (49% APY)", growth_1_year.as_near());
+        println!("  - 2 years: {} NEAR growth (99% total)", growth_2_years.as_near());
+        println!("  - 4 years: {} NEAR growth (199% total)", growth_4_years.as_near());
+    }
+
+    #[test]
+    fn test_incremental_growth_like_real_contract() {
+        // Test incremental growth like the real contract does
+        let config = VenearGrowthConfig::FixedRate(Box::new(VenearGrowthConfigFixedRate {
+            annual_growth_rate_ns: Fraction {
+                numerator: 15854895991882.into(),  // 15,854,895,991,882
+                denominator: 10u128.pow(30).into(), // 10^30
+            },
+        }));
+        
+        let base_balance = NearToken::from_near(100);
+        let one_year_ns = 31_536_000_000_000_000u64;
+        
+        // Calculate growth incrementally like the contract does
+        let start_time: TimestampNs = 0.into();
+        let end_time_1_year: TimestampNs = one_year_ns.into();
+        let end_time_2_years: TimestampNs = (one_year_ns * 2).into();
+        let end_time_3_years: TimestampNs = (one_year_ns * 3).into();
+        let end_time_4_years: TimestampNs = (one_year_ns * 4).into();
+        
+        // Year 1: 0 to 1 year
+        let growth_year_1 = config.calculate(start_time, end_time_1_year, base_balance);
+        println!("Year 1 growth: {} NEAR", growth_year_1.as_near());
+        
+        // Year 2: 1 year to 2 years
+        let growth_year_2 = config.calculate(end_time_1_year, end_time_2_years, base_balance);
+        println!("Year 2 growth: {} NEAR", growth_year_2.as_near());
+        
+        // Year 3: 2 years to 3 years
+        let growth_year_3 = config.calculate(end_time_2_years, end_time_3_years, base_balance);
+        println!("Year 3 growth: {} NEAR", growth_year_3.as_near());
+        
+        // Year 4: 3 years to 4 years
+        let growth_year_4 = config.calculate(end_time_3_years, end_time_4_years, base_balance);
+        println!("Year 4 growth: {} NEAR", growth_year_4.as_near());
+        
+        // Calculate total growth after 2 and 4 years
+        let total_growth_2_years = growth_year_1.as_near() + growth_year_2.as_near();
+        let total_growth_4_years = growth_year_1.as_near() + growth_year_2.as_near() + growth_year_3.as_near() + growth_year_4.as_near();
+        
+        println!("Total growth after 2 years: {} NEAR", total_growth_2_years);
+        println!("Total growth after 4 years: {} NEAR", total_growth_4_years);
+        
+        // Each year should give approximately the same growth (linear)
+        assert_eq!(growth_year_1.as_near(), growth_year_2.as_near());
+        assert_eq!(growth_year_2.as_near(), growth_year_3.as_near());
+        assert_eq!(growth_year_3.as_near(), growth_year_4.as_near());
+    }
+
+    #[test]
+    fn test_venear_balance_update_function() {
+        use crate::VenearBalance;
+        
+        // Test using the actual VenearBalance::update() function like the contract does
+        // Using a numerator to achieve ~49 NEAR growth per year
+        let config = VenearGrowthConfig::FixedRate(Box::new(VenearGrowthConfigFixedRate {
+            annual_growth_rate_ns: Fraction {
+                numerator: 15854895991882.into(),  // This gives ~49 NEAR growth per year
+                denominator: 10u128.pow(30).into(), // 10^30
+            },
+        }));
+        
+        let base_balance = NearToken::from_near(100);
+        let one_year_ns = 31_536_000_000_000_000u64;
+        
+        let mut venear_balance = VenearBalance::from_near(base_balance);
+        println!("Initial balance: {} NEAR base + {} NEAR extra = {} NEAR total", 
+                 venear_balance.near_balance.as_near(),
+                 venear_balance.extra_venear_balance.as_near(),
+                 venear_balance.near_balance.as_near() + venear_balance.extra_venear_balance.as_near());
+        
+        // Year 1: Update from 0 to 1 year (user claiming after 1 year)
+        let start_time: TimestampNs = 0.into();
+        let end_time_1_year: TimestampNs = one_year_ns.into();
+        
+        venear_balance.update(start_time, end_time_1_year, &config);
+        println!("After 1 year: {} NEAR base + {} NEAR extra = {} NEAR total", 
+                 venear_balance.near_balance.as_near(),
+                 venear_balance.extra_venear_balance.as_near(),
+                 venear_balance.near_balance.as_near() + venear_balance.extra_venear_balance.as_near());
+        
+        assert_eq!(venear_balance.near_balance.as_near(), 100); // Base never changes
+        assert!(venear_balance.extra_venear_balance.as_near() > 48 && venear_balance.extra_venear_balance.as_near() <= 50, 
+                "Year 1: extra_venear_balance should be > 48 and <= 50, got {}", 
+                venear_balance.extra_venear_balance.as_near());
+        
+        // Year 2: Update from 1 year to 2 years (user claiming again after another year)
+        let end_time_2_years: TimestampNs = (one_year_ns * 2).into();
+        venear_balance.update(end_time_1_year, end_time_2_years, &config);
+        println!("After 2 years: {} NEAR base + {} NEAR extra = {} NEAR total", 
+                 venear_balance.near_balance.as_near(),
+                 venear_balance.extra_venear_balance.as_near(),
+                 venear_balance.near_balance.as_near() + venear_balance.extra_venear_balance.as_near());
+        
+        // Year 2 assertions
+        assert_eq!(venear_balance.near_balance.as_near(), 100);
+        assert!(venear_balance.extra_venear_balance.as_near() > 98 && venear_balance.extra_venear_balance.as_near() <= 100, 
+                "Year 2: extra_venear_balance should be > 98 and <= 100, got {}", 
+                venear_balance.extra_venear_balance.as_near());
+        
+        // Year 3: Update from 2 years to 3 years
+        let end_time_3_years: TimestampNs = (one_year_ns * 3).into();
+        venear_balance.update(end_time_2_years, end_time_3_years, &config);
+        println!("After 3 years: {} NEAR base + {} NEAR extra = {} NEAR total", 
+                 venear_balance.near_balance.as_near(),
+                 venear_balance.extra_venear_balance.as_near(),
+                 venear_balance.near_balance.as_near() + venear_balance.extra_venear_balance.as_near());
+        
+        assert_eq!(venear_balance.near_balance.as_near(), 100); 
+        assert!(venear_balance.extra_venear_balance.as_near() > 148 && venear_balance.extra_venear_balance.as_near() <= 150, 
+                "Year 3: extra_venear_balance should be > 148 and <= 150, got {}", 
+                venear_balance.extra_venear_balance.as_near());
+        
+        // Year 4: Update from 3 years to 4 years
+        let end_time_4_years: TimestampNs = (one_year_ns * 4).into();
+        venear_balance.update(end_time_3_years, end_time_4_years, &config);
+        println!("After 4 years: {} NEAR base + {} NEAR extra = {} NEAR total", 
+                 venear_balance.near_balance.as_near(),
+                 venear_balance.extra_venear_balance.as_near(),
+                 venear_balance.near_balance.as_near() + venear_balance.extra_venear_balance.as_near());
+        
+        assert_eq!(venear_balance.near_balance.as_near(), 100); 
+        // Note: Due to precision/rounding in the growth calculation, we get 199 instead of 196
+        // This is expected behavior - the growth accumulates with small rounding differences
+        assert!(venear_balance.extra_venear_balance.as_near() > 195 && venear_balance.extra_venear_balance.as_near() <= 200, 
+                "Year 4: extra_venear_balance should be > 195 and <= 200, got {}", 
+                venear_balance.extra_venear_balance.as_near());
+        
+        let total_ve_near = venear_balance.near_balance.as_near() + venear_balance.extra_venear_balance.as_near();
+        
+        println!("   - Base NEAR: {} (never changes)", venear_balance.near_balance.as_near());
+        println!("   - Extra veNEAR: {} (accumulated growth)", venear_balance.extra_venear_balance.as_near());
+        println!("   - Total veNEAR: {} (base + growth)", total_ve_near);
+    }
+}
